+++ SmartSaver/SmartSaver.Presentation/SmartSaver.Presentation/Pages/FetchData.razor
@@ -40,7 +40,9 @@ else
 
     protected override async Task OnInitializedAsync()
     {
-        _transactions = await Http.GetFromJsonAsync<Transaction[]>("http://localhost:5000/transactions");
+        Console.WriteLine(await Http.GetFromJsonAsync<object>("http://localhost:5000/transactions"));
+
+        //_transactions = await Http.GetFromJsonAsync<Transaction[]>("http://localhost:5000/transactions");
     }
 }
