--- SmartSaver/SmartSaver.Presentation/SmartSaver.Presentation/Toasts/Toast.razor
@@ -1,11 +1,0 @@
-﻿@inherits ToastBase
-
-<div class="toast @(IsVisible ? "toast-visible" : null) @BackgroundCssClass">
-    <div class="toast-icon">
-        <i class="fa fa-@IconCssClass" aria-hidden="true"></i>
-    </div>
-    <div class="toast-body">
-        <h5>@Heading</h5>
-        <p>@Message</p>
-    </div>
-</div>
