--- SmartSaver/SmartSaver.Presentation/SmartSaver.Presentation/Toasts/ToastLevel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace SmartSaver.Presentation.Toasts
-{
-    public enum ToastLevel
-    {
-        Info,
-        Success,
-        Warning,
-        Error
-    }
-}
