+++ SmartSaver/SmartSaver.Presentation/SmartSaver.Presentation/wwwroot/css/app.css
@@ -184,7 +184,7 @@ app {
 
 }
 
-
+/*Styled form*/
 .styled-form {
     max-width: 450px;
     font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
