+++ SmartSaver/SmartSaver.Server/SmartSaver.Domain/Models/SavingGoal.cs
@@ -21,7 +21,7 @@ namespace SmartSaver.Domain.Models
 
         public Guid UserId { get; set; }
 
-        public virtual User User { get; set; }
+        //public virtual User User { get; set; }
 
         [NotMapped]
         public double ProgressPercentageValue {
